Self destruct and finishing touches to movements
Finished self destruct functionality. The problem in the previous version was that the range of cells the program was checking was 1 short, which made it only check the upper left 3 cells instead of the 8 cells around the unit.

Rewrote some code for other types of movement actions, as there were some functionalities that weren't working properly.