Attack action refactor and repair action fix
Refactored the code from the attack action because we did not realize that you could use the mod_health function from the game class instead of the unit class. This allows us to skip the remove_dead step.

The health action had a minor problem where instead of healing the other unit, it would damage itself instead.